+++ egui/src/placer.rs
@@ -102,7 +102,7 @@ impl Placer {
     /// Returns where to put the next widget that is of the given size.
     /// The returned `frame_rect` will always be justified along the cross axis.
     /// This is what you then pass to `advance_after_rects`.
-    /// Use `justify_or_align` to get the inner `widget_rect`.
+    /// Use `justify_and_align` to get the inner `widget_rect`.
     pub(crate) fn next_space(&self, child_size: Vec2, item_spacing: Vec2) -> Rect {
         if let Some(grid) = &self.grid {
             grid.next_cell(self.region.cursor, child_size)
@@ -113,11 +113,11 @@ impl Placer {
     }
 
     /// Apply justify or alignment after calling `next_space`.
-    pub(crate) fn justify_or_align(&self, rect: Rect, child_size: Vec2) -> Rect {
+    pub(crate) fn justify_and_align(&self, rect: Rect, child_size: Vec2) -> Rect {
         if let Some(grid) = &self.grid {
-            grid.justify_or_align(rect, child_size)
+            grid.justify_and_align(rect, child_size)
         } else {
-            self.layout.justify_or_align(rect, child_size)
+            self.layout.justify_and_align(rect, child_size)
         }
     }